Create 3__.c
In a company, an employee is paid as under. If his basic salary is less than ₹1500 then HRA = 10% of his basic salary and DA = 90% of his basic salary. if his salary is either equal to or above ₹1500, then HRA = ₹500 and DA = 98% of basic salary. If the employee's salary is input through the keyboard write a program to find his gross salary.