Create 4_D_b.c
In the digital world, colors are specified in Red-Green-Blue (RGB) format, with values of R, G, and B varying on an integer scale from 0 to 255. In print publishing, the colors are mentioned in Cyan-Magenta-Yellow-Black (CMYK) format, with values of C, M, Y, and K varying on a real scale from 0.0 to 1.0. Write a program that converts RGB color to CMYK color as per the following formulae: 
white = max(red/255,green/255,blue/255)
cyan=((white-red/255)/white)
magenta=((white-green/255)/white)
yellow=((white-blue/255)/white)
black=1-white
Note that if the RGB values are all 0, then the CMY values are all 0 and the K value is 1.