Create 5_B_b.c
Write a program to print out all Armstrong numbers between 1 and 500. If the sum of cubes of each digit of the number is equal to the number itself, then the number is called Armstrong. For example,153=(1*1*1)+(5*5*5)+(3*3*3).
Write a program to print out all Armstrong numbers between 1 and 9999. If the sum of power 4 of each digit of the number is equal to the number itself, then the number is called Armstrong.